requirements.txt
requests
python-jose
six
pytest
